post dir change, pre pre build
