+++ mobile-app/src/app/(app)/(tabs)/_layout.tsx
