+++ mobile-app/src/app/(app)/(tabs)/friends.tsx
