+++ mobile-app/src/app/(app)/(tabs)/index.tsx
