+++ mobile-app/src/app/(app)/habits/icon-button.tsx
