+++ mobile-app/src/app/(app)/modals/addfriends.tsx
