+++ mobile-app/src/app/(app)/modals/viewhabit.tsx
