+++ mobile-app/src/app/(app)/specificsettings/widgetsettings.tsx
