+++ mobile-app/src/app/(app)/viewhabit.tsx
